requirements.txt
asgiref==3.7.2
Django==4.2.4
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-filter==23.2
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-phonenumber-field==7.1.0
django-resized==1.0.2
django-rest-registration==0.8.2
djangorestframework==3.14.0
drf-writable-nested==0.7.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
phonenumbers==8.13.19
phonenumberslite==8.13.35
Pillow==10.0.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
